+++ libopendmi/include/opendmi/module/intel.h
@@ -22,7 +22,7 @@ typedef enum dmi_intel_type
     DMI_TYPE_INTEL_RSD_STORAGE_DEVICE      = 194, ///< Intel RSD Storage device information
     DMI_TYPE_INTEL_RSD_TPM                 = 195, ///< Intel RSD TPM information
     DMI_TYPE_INTEL_RSD_TXT                 = 196, ///< Intel RSD TXT information
-    DMI_TYPE_INTEL_RSD_MEMORY_DEVICE_EX    = 197, ///< Intel RSD Memory device extended information
+    DMI_TYPE_INTEL_RSD_MEMORY_DEVICE       = 197, ///< Intel RSD Memory device extended information
     DMI_TYPE_INTEL_RSD_FPGA                = 198, ///< Intel RSD FPGA information
     DMI_TYPE_INTEL_RSD_CABLED_PCIE         = 199, ///< Intel RSD Cabled PCIe port information
     DMI_TYPE_INTEL_RSD_PHYS_DEVICE_MAPPING = 200  ///< Intel RSD SMBIOS physical device mapping